% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/medianTest.R
\name{medianTest}
\alias{medianTest}
\title{Monte Carlo Median test}
\usage{
medianTest(x, fac, n = 999, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{fac}{a factor defining the two groups}

\item{n}{the number of permutations}

\item{...}{other arguments passed to \code{as.randtest}}
}
\description{
This function tests the equality of medians of two groups of observations.
The statistic used is the difference between the medians of the two groups: median(group1) - median(group2).
}
\examples{
## simulate data
set.seed(1)
titer <- c(rexp(100, rate=0.002), rexp(20, rate=0.0005), # control
       rexp(55, rate=0.006), rexp(15, rate=0.0005)) # treatment

type <- factor(rep(c("control","treatment"), c(120,70)))

VL <- data.frame(titer=titer, type=type)

## plot data
if(require(ggplot2)){
ggplot(VL, aes(x=titer)) + geom_density(aes(fill=type), alpha=.5, adjust=4)
}

## compare t.test and median test
t.test(titer~type)

medianTest(titer,type)
}

